+++ crates/vm/levm/src/db/cache.rs
@@ -3,6 +3,10 @@ use std::collections::HashMap;
 
 pub type CacheDB = HashMap<Address, Account>;
 
+pub fn account_is_cached(cached_accounts: &CacheDB, address: &Address) -> bool {
+    cached_accounts.contains_key(address)
+}
+
 pub fn get_account<'cache>(
     cached_accounts: &'cache CacheDB,
     address: &Address,